+++ agent-loader/pom.xml
@@ -1,43 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (C) 2015 Electronic Arts Inc.  All rights reserved.
 
-  Redistribution and use in source and binary forms, with or without
-  modification, are permitted provided that the following conditions
-  are met:
-
-  1.  Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-  2.  Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in the
-      documentation and/or other materials provided with the distribution.
-  3.  Neither the name of Electronic Arts, Inc. ("EA") nor the names of
-      its contributors may be used to endorse or promote products derived
-      from this software without specific prior written permission.
-
-  THIS SOFTWARE IS PROVIDED BY ELECTRONIC ARTS AND ITS CONTRIBUTORS "AS IS" AND ANY
-  EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-  DISCLAIMED. IN NO EVENT SHALL ELECTRONIC ARTS OR ITS CONTRIBUTORS BE LIABLE FOR ANY
-  DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-  (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-  ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
-  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-  -->
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>com.ea.agentloader</groupId>
-        <artifactId>ea-agent-loader-parent</artifactId>
-        <version>1.0.4-SNAPSHOT</version>
+        <groupId>com.linkedin.agentloader</groupId>
+        <artifactId>linkedin-agent-loader-parent</artifactId>
+        <version>1.0.0</version>
         <relativePath>..</relativePath>
     </parent>
-    <artifactId>ea-agent-loader</artifactId>
-    <name>EA Agent Loader</name>
+    <artifactId>linkedin-agent-loader</artifactId>
+    <name>Linkedin Agent Loader</name>
 
     <build>
         <plugins>
