+++ agent-loader/src/main/resources/com/linkedin/agentloader/shaded/AttachOperationFailedException.class
