+++ agent-loader/src/main/resources/com/linkedin/agentloader/shaded/AttachPermission.class
