+++ agent-loader/src/main/resources/com/linkedin/agentloader/shaded/HotSpotAttachProvider$HotSpotVirtualMachineDescriptor.class
