+++ agent-loader/src/main/resources/com/linkedin/agentloader/shaded/VirtualMachine.class
